Refactored Store class and Locate class to simulated GC indexing package
