+++ DFS/src/dfsUfsCore/dfs3Mgr/WatchDir.java
@@ -1,4 +1,4 @@
-package dfsUfsCore.dfsMgr;
+package dfsUfsCore.dfs3Mgr;
 
 import java.nio.file.*;
 import static java.nio.file.StandardWatchEventKinds.*;
