+++ DFS/src/dfsUfsCore/dfs3xmlHandler/ReadInode.java
@@ -1,10 +1,8 @@
-package dfsUfsCore.xmlHandler;
-import java.nio.charset.StandardCharsets;
+package dfsUfsCore.dfs3xmlHandler;
 import java.security.KeyFactory;
 import java.security.NoSuchAlgorithmException;
 import java.security.PublicKey;
 import java.security.spec.InvalidKeySpecException;
-import java.security.spec.KeySpec;
 import java.security.spec.X509EncodedKeySpec;
 import java.util.*;
