+++ DFS/src/dfsUfsCore/dfs3xmlHandler/WriteObject.java
@@ -1,4 +1,4 @@
-package dfsUfsCore.xmlHandler;
+package dfsUfsCore.dfs3xmlHandler;
 
 import java.util.Base64;
