+++ DFS/src/simulateGC/communication/Receiver.java
@@ -1,7 +1,7 @@
 package simulateGC.communication;
 
-import dfsUfsCore.dfsMgr.DFS3BufferMgr;
-import dfsUfsCore.xmlHandler.XMLReader;
+import dfsUfsCore.dfs3Mgr.DFS3BufferMgr;
+import dfsUfsCore.dfs3xmlHandler.XMLReader;
 
 import java.io.*;
 import java.net.ServerSocket;
